+++ app/models/result.rb
@@ -37,6 +37,44 @@ class Result < ApplicationRecord
     end
   end
 
+  # Release or unrelease the submissions of a set of groupings.
+  def self.set_release_on_results(grouping_ids, release)
+    groupings = Grouping.where(id: grouping_ids)
+    without_submissions = groupings.where.not(id: groupings.joins(:current_submission_used))
+
+    if without_submissions.present?
+      group_names = without_submissions.joins(:group).pluck(:group_name).join(', ')
+      raise StandardError, I18n.t('submissions.errors.no_submission', group_name: group_names)
+    end
+
+    without_complete_result = groupings.joins(:current_result)
+                                       .where.not('results.marking_state': Result::MARKING_STATES[:complete])
+
+    if without_complete_result.present?
+      group_names = without_complete_result.joins(:group).pluck(:group_name).join(', ')
+      if release
+        raise StandardError, I18n.t('submissions.errors.not_complete', group_name: group_names)
+      else
+        raise StandardError, I18n.t('submissions.errors.not_complete_unrelease', group_name: group_names)
+      end
+    end
+
+    result = Result.where(id: groupings.joins(:current_result).pluck('results.id'))
+                   .update_all(released_to_students: release)
+
+    if release
+      groupings.includes(:accepted_students).find_each do |grouping|
+        grouping.accepted_students.each do |student|
+          if student.receives_results_emails?
+            NotificationMailer.with(user: student, grouping: grouping).release_email.deliver_later
+          end
+        end
+      end
+    end
+
+    result
+  end
+
   # Calculate the total mark for this submission
   def get_total_mark
     user_visibility = is_a_review? ? :peer_visible : :ta_visible